Fix proxying; added version. Read on.
Turns out that nim does not take empty strings as an empty proxy.
This has been fixed, so that no proxy is set if none is found.

A --version flag was also added. It's nice to be proper.

The rest is just some odds and ends cleaning up and some
more complete ad-hoc tests.